+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **DecaLog** are documented in this *changelog*.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased - Will be 2.0.0]
+### Added
+- New WP-CLI command to manage loggers (list, start, pause, clean, purge and remove).
 ### Changed
 - The PHP listener now takes care of activated/deactivated modules between web server and command-line configurations.
 ### Fixed